pyproject.toml
[project]
name = "jointoit-testtask"
version = "0.1.0"
description = "Django REST-Api for Event Management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.3",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
    "redis>=6.2.0",
]
